Added handling of list of timestamp
In GelfTcp there is already the possibility to have list of messages, i've added also list of timestamp so that, only in case of list of messages, they can be sent each with their message.
Is also possible to have a list  of messages with only 1 timestamp so also that possibility is handled.

In GelfHttp is a bit more complicated.
There was not a list handling but, according to graylog documentation, they introduced a way to handle bulk http GELF post.
https://go2docs.graylog.org/5-0/getting_in_log_data/ingest_gelf.html
So i've added the same list handling from Tcp but, as the bulking is made appending dumped JSON with \n\r at the end, and, most important, must be avoided to have \n or \r inside the dumped messages, i had to make some cases that handle all the above.
To keep the rest as it is i've jumped the single message before so only 1 variable needs to be handled.